Improved makefile, unit tests framework, added mods.lsp, ...
The makefiles have been improved again and changes make to the interpreter to
make it more compatible with Windows, the library still compiles perfectly
under Linux.

Two new directories have been added to the rpath of the "lisp" executable that
is built, this has not been tested under Windows. It adds the current and
"mod/" directories to the shared library search path.

A unit test framework has been added, it has uncovered some assumptions I have
made in the code already. The "util.c" nearly has full coverage, it might be
quite difficult to test the interpreter itself sensibly however.

I have found a bug in "read.c" and how it interacts with the rest of the
interpreter when it fails to read in a string correctly, this will be
documented and then fixed, it produces quite confusing behavior.

As usual, there are other minor changes throughout.

Also of note, files in the "mod/lisp" directory are not been added with
the "git add ." command, this should not happen, although the entry "lisp" is
in the ".gitignore" file, (see <https://git-scm.com/docs/gitignore>).

+++ liblisp.h
@@ -195,9 +195,11 @@ LIBLISP_API char *vstrcatsep(const char *separator, const char *first, ...);
 /** @brief  Calculate the binary logarithm, for more efficient examples see
  *          http://graphics.stanford.edu/~seander/bithacks.html or
  *          "Bit Twiddling Hacks by Sean Eron Anderson"
+ *          This version returns INT32_MIN on an error (which cannot
+ *          occur normally).
  *  @param  v         Value to calculate the binary logarithm of
  *  @return uint8_t   Binary log**/
-LIBLISP_API uint8_t ilog2(uint64_t v);
+LIBLISP_API int32_t ilog2(uint64_t v);
 
 /** @brief  Calculate an integer exponentiation.
  *  From:
@@ -606,7 +608,10 @@ LIBLISP_API int is_number(const char *buf);
 /**@brief  this is much like is_number, but for floating point numbers
  *         matches "[-+]?[0-9]*\.?[0-9]+([eE][-+]?[0-9]+)?", and
  *         excludes valid conversions added in c99 (such as "inf" or
- *         "nan").
+ *         "nan"). This is not a test that a number is exclusively
+ *         a floating point number, only that is can be converted into
+ *         one sensible, to test for that use is_number() on the
+ *         string first.
  * @param  buf ascii delimited string to check if it is a float
  * @return 0 if it is not a float, non zero if it is**/
 LIBLISP_API int is_fnumber(const char *buf);
